Init
The sbp class and two functions : sbp_include include a sbp file or the php file if it's up to date or if the sbp file does not exists. If any of them exist, sbp_include throw an exception. sbp_include_if_exists do the same but just return false in case of error and does not throw any exception.